refactor: isolate tag team employment validation
